clean: check hkcu:console exists before cleaning children

+++ libexec/clean.ps1
@@ -5,4 +5,7 @@
 # Background: The Console will save program-specific overrides to the registry when you
 # edit the console properties directly. These overrides can prevent your concfg settings
 # from working properly.
-gci hkcu:console | % { rm "registry::$($_.name)" }
+
+if(test-path hkcu:console) {
+    gci hkcu:console | % { rm "registry::$($_.name)" }
+}
